Add optional center of mass adjustment and theoretical shift comparison
Introduce `use_theoretical_shift` parameter to `calculate_shape` method, enabling optional adjustment to place the center of mass at the origin or rely on the theoretical z_shift. Add functionality for calculating center of mass and visualizing both theoretical and calculated shifts in the plot. Update labels and shift-related information in the UI display.

Signed-off-by: AleksanderAugustyn <aleksander.augustyn@ncbj.gov.pl>